chore: fixes the types and moves files arounf
