+++ src/core_modules/capture-core/components/LockedSelector/QuickSelector/OrgUnitSelector/orgUnitList.epics.js
@@ -10,7 +10,7 @@ import {
     setSearchRoots,
     setSearchRootsFailed,
     showLoadingIndicator,
-} from '../actions/orgUnitList.actions';
+} from '../OrgUnitSelector/orgUnitList.actions';
 import { set as setStoreRoots } from '../../../FormFields/New/Fields/OrgUnitField/orgUnitRoots.store';
 import { LOADING_INDICATOR_TIMEOUT } from '../../../../constants';
