+++ src/core_modules/capture-core/components/LockedSelector/QuickSelector/ProgramSelector/programSelector.css
