--- src/core_modules/capture-core/components/LockedSelector/QuickSelector/programSelector.css
@@ -1,4 +1,0 @@
-#program-selector .Select-control {
-    height: 46px;
-    line-height: 46px;    
-}
