+++ src/core_modules/capture-core/reducers/descriptions/organisationUnits.reducerDescription.js
@@ -2,7 +2,7 @@
 import { createReducerDescription } from '../../trackerRedux/trackerReducer';
 import { actionTypes as editEventActionTypes } from '../../components/Pages/EditEvent/editEvent.actions';
 import { actionTypes as viewEventActionTypes } from '../../components/Pages/ViewEvent/ViewEventComponent/viewEvent.actions';
-import { actionTypes as setOrgUnitActionTypes } from '../../components/LockedSelector/QuickSelector/actions/QuickSelector.actions';
+import { actionTypes as setOrgUnitActionTypes } from '../../components/LockedSelector/QuickSelector/QuickSelector.actions';
 import { lockedSelectorActionTypes } from '../../components/LockedSelector/LockedSelector.actions';
 import { orgUnitListActionTypes } from '../../components/LockedSelector/QuickSelector';
 import { set as setStoreRoots } from '../../components/FormFields/New/Fields/OrgUnitField/orgUnitRoots.store';
